Added CRUD to AddressController and separated each controller into category.
